packages/dotmac-platform-services/pyproject.toml
[build-system]
requires = ["poetry-core"]

build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dotmac-platform-services"
version = "1.0.0"
description = "DotMac Platform Services - Unified auth, secrets, and observability"
authors = ["DotMac Team <dev@dotmac.com>"]

readme = "README.md"
packages = [{include = "dotmac", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.110.0"
pydantic = ">=2.5.0"
sqlalchemy = ">=2.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
hvac = "^2.0.0"
prometheus-client = "^0.19.0"
opentelemetry-api = ">=1.21.0"
opentelemetry-sdk = ">=1.21.0"
opentelemetry-instrumentation-fastapi = ">=0.42b0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"

[tool.ruff]
line-length = 100
target-version = "py39"
extend-exclude = ["migrations/"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "BLE", "B", "A", "C4", "ICN", "PIE", "PT", "RSE", "RET", "SIM", "ARG", "PTH", "ERA", "PL", "RUF"]

ignore = ["ANN101", "ANN102", "S101", "PLR0913", "PLR0915"]

[tool.mypy]
python_version = "3.9"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest]
[tool.pytest.ini_options]
testpaths = ["tests"]

python_files = ["test_*.py"]

python_classes = ["Test*"]

python_functions = ["test_*"]

asyncio_mode = "auto"
markers = ["unit: Unit tests", "integration: Integration tests", "e2e: End-to-end tests", "slow: Slow running tests", "auth: Authentication related tests", "observability: Observability related tests", "secrets: Secrets management tests"]

addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/dotmac",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=60"
]
